Move core code from __main__ to _maxjob, while import all of it in both __main__ and __init__.
