+++ maxjob/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# :coding: utf-8
+
+"""maxjob unclenches maxscript execution with 3ds Max."""
+
+from maxjob._maxjob import *
